requirements.txt
ipywidgets==8.0.4
matplotlib==3.6.3
numpy==1.23.5
pandas==1.5.3
torch==2.0.1
tqdm==4.65.0
bs4==0.0.1
jsonlines==3.1.0
pytest-playwright==0.3.2
trafilatura==1.6.1
tiktoken==0.3.0
bert-extractive-summarizer==0.10.1
sentence-transformers==2.2.2
markdown==3.4.1
nltk==3.8.1
numba==0.57.1
Flask==2.2.5
flask-cors==4.0.0
gunicorn==20.1.0
markdownify==0.11.6
python-dotenv==1.0.0
psycopg==3.1.8
pgvector==0.1.6
langchain==0.0.230
dash==2.9.1
bertopic==0.14.1
cleanlab==2.3.0

alembic==1.11.1
attrs==23.1.0
certifi==2023.5.7
chardet==5.1.0
click==8.1.4
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
dash-bootstrap-components==1.3.1
decorator==5.1.1
dtale==3.2.0
Flask-Compress==1.13.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
future==0.18.3
idna==3.4
ipython-genutils==0.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.1
jupyter-core==5.3.1
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
nbformat==5.9.1
plotly==5.15.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2023.3
requests==2.31.0
retrying==1.3.4
six==1.16.0
SQLAlchemy==2.0.18
traitlets==5.9.0
urllib3==2.0.3
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1